fix : 타입 수정
