+++ admin/src/type/common/type.ts
@@ -1,0 +1,5 @@
+export interface Response {
+  isSuccess: boolean;
+  code: string;
+  message: string;
+}
