--- admin/src/type/eventManagement/eventType.ts
@@ -1,10 +1,0 @@
-export interface FcFsRequest {
-  startDate: string;
-  endDate: string;
-  startTime: string; // (pattern = "HH:mm:ss")
-}
-
-export interface DrawRequest {
-  startTime: string;
-  endTime: string;
-}
